PatternGenerator configured for generating 2D distributions phiB vs phiDist
GoldenPatternWithStat.h StatArrayType insing unsigned short - remember
to change it for int for the pattern generation

PatternOptimizerBase.cc the 2D hists are sotred only for the refLayer 0
and 2

CandidateSimMuonMatcher.cc requires fixes of the simMu eta cuts

+++ L1Trigger/L1TMuonOverlapPhase1/interface/Omtf/GoldenPatternWithStat.h
@@ -17,7 +17,7 @@ class OMTFConfiguration;
 class GoldenPatternWithStat : public GoldenPatternWithThresh {
 public:
   static const unsigned int STAT_BINS = 1;  //TODO change to 4!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-  typedef boost::multi_array<int, 4> StatArrayType;
+  typedef boost::multi_array<unsigned short, 4> StatArrayType;
   //GoldenPatternWithStat(const Key & aKey) : GoldenPattern(aKey) {}
 
   GoldenPatternWithStat(const Key& aKey, unsigned int nLayers, unsigned int nRefLayers, unsigned int nPdfAddrBits);
